Fix Feature Flag Slack notifications
